Update ParticipantService to surround UploadDao and ParticipantDao calls within TransactionScope
