+++ .gitignore
@@ -7,6 +7,8 @@
 !.vscode/extensions.json
 *.code-workspace
 
+.vs/*
+
 # Local History for Visual Studio Code
 .history/
 iac/env/tts/eth.bash
